+++ packages/ubcli/lib/ddlGenerators/db/PostgreSQL.js
@@ -14,8 +14,7 @@ const DBAbstract = require('./DBAbstract')
 const DBPostgresPolicies = {
   PolicyTypes: {
     tenantOnly: `(mi_tenantid = (current_setting('ub.tenantID'::text))::bigint)`,
-    tenantAndCommon:`(mi_tenantid = ANY (ARRAY[(0)::bigint, (current_setting('ub.tenantID'::text))::bigint]))`,
-    tenantCommonAndSystem: `(mi_tenantid = ANY (ARRAY[(0)::bigint, (1)::bigint, (current_setting('ub.tenantID'::text))::bigint]))`
+    tenantAndCommon:`(mi_tenantid = ANY (ARRAY[(0)::bigint, (current_setting('ub.tenantID'::text))::bigint]))`
   },
 
   /**
@@ -280,10 +279,10 @@ left join pg_namespace n on n.oid = c.relnamespace
         data: policiesSQL,
         URLParams: { CONNECTION: this.dbConnectionConfig.name }
       })
-      for (const /** @type {{policyName: string, text: string}} */ policyFromDb of policiesFromDb) {
-        asIsTable.addOther({
+      for (const /** @type {object} */ policyFromDb of policiesFromDb) {
+        asIsTable.addPolicy({
           type: DBPostgresPolicies.getPolicyType(policyFromDb),
-          name: 'policy$' + policyFromDb.name
+          name: policyFromDb.name
         })
       }
 
@@ -445,34 +444,29 @@ left join pg_namespace n on n.oid = c.relnamespace
     this.DDL.createTable.statements.push(res.join(''))
   }
 
-  /** @override */
-  genCodeEnableMultitenancy (table) {
+  genCodeEnableRls (table) {
     this.DDL.others.statements.push(
       `ALTER TABLE ${table.name} ENABLE ROW LEVEL SECURITY`
     )
-    if (table.name === 'uba_subject') {
-      // The "uba_subject" is a union (the "unity" mixin) from "uba_user",
-      // "uba_group" and "uba_role" entities.
-      // The problematic part is "uba_role" records, because "uba_role" is the only entity, which does not
-      // contain "tenants" mixin and, therefore, does not contain mi_tenantID.
-      // Records in "uba_subject" for roles have mi_tenantID=0.
-      this.DDL.others.statements.push(
-        `CREATE POLICY ${table.name}_policy ON ${table.name} USING ${DBPostgresPolicies.PolicyTypes.tenantAndCommon}`
-      )
-    } else {
-      this.DDL.others.statements.push(
-        `CREATE POLICY ${table.name}_policy ON ${table.name} USING ${DBPostgresPolicies.PolicyTypes.tenantOnly})`
-      )
-    }
   }
 
-  /** @override */
-  genCodeDisableMultitenancy (table) {
+  genCodeDisableRls (table) {
     this.DDL.others.statements.push(
       `ALTER TABLE ${table.name} DISABLE ROW LEVEL SECURITY`
     )
+  }
+
+  genCodeDropPolicy (table, policy) {
     this.DDL.others.statements.push(
-      `DROP POLICY IF EXISTS ${table.name}_policy ON ${table.name}`
+      `DROP POLICY IF EXISTS ${policy.name} ON ${table.name}`
+    )
+  }
+
+  genCodeCreateOrAlterPolicy (table, policy, isAlter) {
+    const mainSQL = isAlter ? 'ALTER POLICY' : 'CREATE POLICY'
+    const policySQL = DBPostgresPolicies.PolicyTypes[policy.type]
+    this.DDL.others.statements.push(
+      `${mainSQL} ${policy.name} ON ${table.name} USING ${policySQL}`
     )
   }
 
@@ -622,7 +616,7 @@ left join pg_namespace n on n.oid = c.relnamespace
    * @param {number} len
    * @param {number}  prec
    * @param {number}  scale
-   * @return {String}
+   * @return {never}
    */
   dataBaseTypeToUni (dataType, len, prec, scale) {
     dataType = dataType.toUpperCase()