+++ packages/ubcli/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- Multitenancy: instead of enable/disable multitenancy,
+  compare RLS policies on tables and compare has/does not have RLS enable flag on-off.
 
 ### Deprecated